refactor: #18 밋팀 정보 Mock/상수 데이터 따로 파일로 저장
